// src/generated/css/StyleSheet.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';
import type { DesignToken } from './DesignToken';
import type { Theme } from './Theme';
import type { CssStyle } from './CssStyle';
import type { StyleRule } from './StyleRule';

/**
 * StyleSheet
 * @generated
 */
export interface StyleSheet extends EObject {
  name?: string;
  tokens: DesignToken[];
  themes: Theme[];
  styles: CssStyle[];
  rules: StyleRule[];

}